+++ classes/Hooks.php
@@ -18,27 +18,21 @@ class Hooks
      * Support share print for modules
      *
      * @param \ModuleModel $objModel
-     * @param string       $strBuffer
-     * @param \Module      $objModule
+     * @param string $strBuffer
+     * @param \Module $objModule
      *
      * @return string
      */
     public static function getFrontendModuleHook(\ModuleModel $objModel, $strBuffer, \Module $objModule)
     {
-        if (!$objModel->addShare)
-        {
+        if (!$objModel->addShare) {
             return $strBuffer;
-        }
-        else {
-            if (Request::hasGet(Share::SHARE_REQUEST_PARAMETER_PRINT))
-            {
+        } else {
+            if (Request::hasGet(Share::SHARE_REQUEST_PARAMETER_PRINT)) {
                 return Share::renderPrintableModule($objModel, $strBuffer, $objModule);
-            }
-            elseif (Request::hasGet(Share::SHARE_REQUEST_PARAMETER_PDF))
-            {
+            } elseif (Request::hasGet(Share::SHARE_REQUEST_PARAMETER_PDF)) {
                 return Share::renderPDFModule($objModel, $strBuffer, $objModule);
-            }
-            else {
+            } else {
                 return $strBuffer;
             }
         }